+++ frontend-share/src/instrumentation-client.ts
@@ -1,0 +1,7 @@
+import * as Sentry from '@sentry/nextjs';
+
+import { getSentryInitOptions } from './lib/sentry';
+
+Sentry.init(getSentryInitOptions());
+
+export const onRouterTransitionStart = Sentry.captureRouterTransitionStart;